Update Learn More Club Button for Signed In
Update the Learn More page "Start a Club Now" button to change based on
whether the User is logged in. If the User is logged in, it reads "My
Club" and directs the User to their edit Club path. If the User is not
logged in, it assumes it is a new User and presents "Start a Club Now"
with a redirect to the User registration page.